ajax_system/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-cors==4.0.0
python-docx==1.1.1